+++ src/datumaro/util/image.py
@@ -1,6 +1,7 @@
 # Copyright (C) 2019-2023 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
+from __future__ import annotations
 
 import importlib
 import os
@@ -8,9 +9,10 @@ import os.path as osp
 import shlex
 import weakref
 from contextlib import contextmanager
+from contextvars import ContextVar
 from enum import Enum, auto
 from io import BytesIO, IOBase
-from typing import Any, Callable, Dict, Iterable, Iterator, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Iterator, Optional, Tuple, Union
 
 import numpy as np
 
@@ -29,19 +31,73 @@ class _IMAGE_BACKENDS(Enum):
     PIL = auto()
 
 
-_IMAGE_BACKEND = None
+_IMAGE_BACKEND: ContextVar[_IMAGE_BACKENDS] = ContextVar("_IMAGE_BACKENDS")
 _image_loading_errors = (FileNotFoundError,)
 try:
     importlib.import_module("cv2")
-    _IMAGE_BACKEND = _IMAGE_BACKENDS.cv2
+    _IMAGE_BACKEND.set(_IMAGE_BACKENDS.cv2)
 except ModuleNotFoundError:
     import PIL
 
-    _IMAGE_BACKEND = _IMAGE_BACKENDS.PIL
+    _IMAGE_BACKEND.set(_IMAGE_BACKENDS.PIL)
     _image_loading_errors = (*_image_loading_errors, PIL.UnidentifiedImageError)
 
 from datumaro.util.image_cache import ImageCache
 from datumaro.util.os_util import find_files
+
+if TYPE_CHECKING:
+    from PIL.Image import Image as PILImage
+
+
+class ImageColorScale(Enum):
+    """Image color scale
+
+    - UNCHANGED: Use the original image's scale (default)
+    - COLOR: Use 3 channels (it can ignore the alpha channel or convert the gray scale image to BGR)
+    """
+
+    UNCHANGED = 0
+    COLOR = 1
+
+    def convert_cv2(self, img: np.ndarray) -> np.ndarray:
+        import cv2
+
+        if self == ImageColorScale.COLOR:
+            return cv2.imdecode(img, cv2.IMREAD_COLOR)
+
+        return cv2.imdecode(img, cv2.IMREAD_UNCHANGED)
+
+    def convert_pil(self, img: PILImage) -> PILImage:
+        if self == ImageColorScale.COLOR:
+            return img.convert("RGB")
+
+        return img
+
+
+IMAGE_COLOR_SCALE: ContextVar[ImageColorScale] = ContextVar(
+    "IMAGE_COLOR_SCALE", default=ImageColorScale.UNCHANGED
+)
+
+
+@contextmanager
+def decode_image_context(color_scale: ImageColorScale):
+    """Change Datumaro image decoding color scale.
+
+    For model training, it is recommended to use this context manager
+    to load images in the BGR 3-channel format. For example,
+
+    .. code-block:: python
+
+        import datumaro as dm
+        with decode_image_context(ImageColorScale.COLOR):
+            item: dm.DatasetItem
+            img_data = item.media_as(dm.Image).data
+            assert img_data.shape[-1] == 3  # It should be a 3-channel image
+    """
+    curr_ctx = IMAGE_COLOR_SCALE.get()
+    IMAGE_COLOR_SCALE.set(color_scale)
+    yield
+    IMAGE_COLOR_SCALE.set(curr_ctx)
 
 
 def load_image(path: str, dtype: DTypeLike = np.uint8, crypter: Crypter = NULL_CRYPTER):
@@ -49,7 +105,7 @@ def load_image(path: str, dtype: DTypeLike = np.uint8, crypter: Crypter = NULL_C
     Reads an image in the HWC Grayscale/BGR(A) [0; 255] format (default dtype is uint8).
     """
 
-    if _IMAGE_BACKEND == _IMAGE_BACKENDS.cv2:
+    if _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.cv2:
         # cv2.imread does not support paths that are not representable
         # in the locale encoding on Windows, so we read the image bytes
         # ourselves.
@@ -58,24 +114,13 @@ def load_image(path: str, dtype: DTypeLike = np.uint8, crypter: Crypter = NULL_C
             image_bytes = crypter.decrypt(f.read())
 
         return decode_image(image_bytes, dtype=dtype)
-    elif _IMAGE_BACKEND == _IMAGE_BACKENDS.PIL:
-        from PIL import Image
+    elif _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.PIL:
+        with open(path, "rb") as f:
+            image_bytes = crypter.decrypt(f.read())
 
-        if not crypter.is_null_crypter:
-            raise DatumaroError("PIL backend should have crypter=NullCrypter.")
+        return decode_image(image_bytes, dtype=dtype)
 
-        image = Image.open(path)
-        image = np.asarray(image, dtype=dtype)
-        if len(image.shape) == 3 and image.shape[2] in {3, 4}:
-            image = np.array(image)  # Release read-only
-            image[:, :, :3] = image[:, :, 2::-1]  # RGB to BGR
-    else:
-        raise NotImplementedError()
-
-    assert len(image.shape) in {2, 3}
-    if len(image.shape) == 3:
-        assert image.shape[2] in {3, 4}
-    return image
+    raise NotImplementedError(_IMAGE_BACKEND)
 
 
 def copyto_image(
@@ -130,7 +175,7 @@ def save_image(
     # NOTE: OpenCV documentation says "If the image format is not supported,
     # the image will be converted to 8-bit unsigned and saved that way".
     # Conversion from np.int32 to np.uint8 is not working properly
-    backend = _IMAGE_BACKEND
+    backend = _IMAGE_BACKEND.get()
     if dtype == np.int32:
         backend = _IMAGE_BACKENDS.PIL
     if backend == _IMAGE_BACKENDS.cv2:
@@ -172,7 +217,7 @@ def encode_image(image: np.ndarray, ext: str, dtype: DTypeLike = np.uint8, **kwa
     if not kwargs:
         kwargs = {}
 
-    if _IMAGE_BACKEND == _IMAGE_BACKENDS.cv2:
+    if _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.cv2:
         import cv2
 
         params = []
@@ -188,7 +233,7 @@ def encode_image(image: np.ndarray, ext: str, dtype: DTypeLike = np.uint8, **kwa
         if not success:
             raise Exception("Failed to encode image to '%s' format" % (ext))
         return result.tobytes()
-    elif _IMAGE_BACKEND == _IMAGE_BACKENDS.PIL:
+    elif _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.PIL:
         from PIL import Image
 
         if ext.startswith("."):
@@ -211,16 +256,17 @@ def encode_image(image: np.ndarray, ext: str, dtype: DTypeLike = np.uint8, **kwa
 
 
 def decode_image(image_bytes: bytes, dtype: DTypeLike = np.uint8) -> np.ndarray:
-    if _IMAGE_BACKEND == _IMAGE_BACKENDS.cv2:
-        import cv2
+    ctx_color_scale = IMAGE_COLOR_SCALE.get()
 
+    if _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.cv2:
         image = np.frombuffer(image_bytes, dtype=np.uint8)
-        image = cv2.imdecode(image, cv2.IMREAD_UNCHANGED)
+        image = ctx_color_scale.convert_cv2(image)
         image = image.astype(dtype)
-    elif _IMAGE_BACKEND == _IMAGE_BACKENDS.PIL:
+    elif _IMAGE_BACKEND.get() == _IMAGE_BACKENDS.PIL:
         from PIL import Image
 
         image = Image.open(BytesIO(image_bytes))
+        image = ctx_color_scale.convert_pil(image)
         image = np.asarray(image, dtype=dtype)
         if len(image.shape) == 3 and image.shape[2] in {3, 4}:
             image = np.array(image)  # Release read-only